rename subcommand
